server performs health checks

+++ scripts/cleanup.sh
@@ -2,5 +2,10 @@
 
 THIS_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )";
 
-rm -rf $THIS_DIR/../download
-rm $THIS_DIR/../imgex.db
+if [ -d $THIS_DIR/../download ]; then
+    rm -rf $THIS_DIR/../download;
+fi;
+
+if [ -f $THIS_DIR/../imgex.db ]; then
+    rm $THIS_DIR/../imgex.db;
+fi;
